IGNITE-25156 Make BinarySchemaRegistry package private
